Remove unused modules and losses
